pyproject.toml
[project]
name = "flame"
dynamic = ["version"]
description = "A minimal training framework for scaling FLA models"
readme = "README.md"
authors = [
    { name = "Songlin Yang", email = "yangsl66@mit.edu" },
    { name = "Yu Zhang", email = "yzhang.cs@outlook.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    'flash-linear-attention',
    'torch>=2.5',
    'torchdata',
    'transformers>=4.45.0',
    'triton>=3.0',
    'datasets>=3.3.0',
    'einops',
    'ninja',
    'wandb',
    'tiktoken',
    'tensorboard',
    'cloudpickle',
]

[project.optional-dependencies]
dev = ["pytest"]

[project.urls]
Homepage = "https://github.com/fla-org/flame"

[build-system]
requires = ["setuptools>=45", "wheel", "ninja", "torch"]

[tool.isort]
line_length = 127
multi_line_output = 3
